+++ xb_social_ai_planner/wizard/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import xb_social_generate_month_wizard
+from . import xb_social_refine_wizard